Break up permissions file.
