+++ homeassistant/auth/permissions/common.py
@@ -1,0 +1,6 @@
+"""Common code for permissions."""
+from typing import (  # noqa: F401
+    Mapping, Union)
+
+CategoryType = Union[Mapping[str, 'CategoryType'], bool, None]
+PolicyType = Mapping[str, CategoryType]
